move Dict39 to lib/datatypes
